==> R/13_lc_introns_to_TSSs_three_database.Rmd <==
---
title: "lc_introns_to_TSSs_three_database"
output: pdf_document
---


Caveat: we can only assign 1st introns if the they are annotated to a transcript - here we add refseq and fantom cat annotations to the ensembl one.  
Also adding in ALL introns to get an accurate background

makes white_TSS, beige_TSS + non_sig_TSS
white_promoters + beige_promoters (unmerged) 

```{r}
library(dplyr)
library(tidyr)
library(ggplot2)
library(readr)
library(here)
here::i_am("R/13_lc_introns_to_TSSs_three_database.Rmd")
```

```{r}
LEAF_INDEX = here('annotations')
leaf_junctions = read.delim(here("31_leafcutter","three_database_info_all_junctions.tsv"))
dim(leaf_junctions)
head(leaf_junctions)
```


```{r}
leaf_junctions = mutate(leaf_junctions, pot_first_intron = min_intron_number == 1, intron_id = paste(chr, start, end, cluster_id,sep=":"))
leaf_junctions$pot_first_intron = factor(leaf_junctions$pot_first_intron, levels=c("TRUE","FALSE"))
leaf_junctions$annotation = factor(leaf_junctions$annotation, levels=c("gencode","refseq","fantom_cat"))

length(unique(leaf_junctions$intron_id))
sig = filter(leaf_junctions, p.adjust < 0.05 & abs(deltapsi)> 0.1)
nrow(sig); #length(unique(sig$intron_id))


summary(sig$min_intron_number)
ggplot(sig, aes(x=min_intron_number, fill=mode_intron_number==1)) + geom_bar()
ggplot(sig, aes(x=min_intron_number==1, fill=min_intron_number==1)) + geom_bar()
head(leaf_junctions)

summary(sig$pot_first_intron)
table(sig$pot_first_intron, sig$annotation)
 ggplot(filter(sig, !is.na(pot_first_intron) ),aes(x=pot_first_intron, fill=pot_first_intron)) + geom_bar()  +
  labs(x="Is first exon junction?", y="Number of exon junctions") +
         theme_bw(base_size=18) + theme(axis.ticks.x = element_blank(), panel.border = element_blank()) 
ggplot(filter(sig, !is.na(pot_first_intron) ),aes(x=annotation, fill=pot_first_intron)) + geom_bar(position="fill")  +
  labs(x="Is first exon junction?", y="Number of exon junctions") +
         theme_bw(base_size=18) + theme(axis.ticks.x = element_blank(), panel.border = element_blank()) 
ggsave(here("R/plots", "annotated_first_intron_barplot.pdf"))
```


## Get 1st exon coords for 1st introns
per intron, could be several 1st exons/TSSs.  but thats okay, just log them here
```{r}
first_introns = filter(leaf_junctions, pot_first_intron==TRUE) %>% rename(intron_start = start, intron_end = end)
nrow(first_introns)
nrow(filter(first_introns, p.adjust < 0.05 & abs(deltapsi)>0.1))#471 1st introns
head(first_introns)
table(first_introns$strand)
```


```{r}
prefixes = c("gencode/gencode.v32", "refseq/refseq_2023","fantom/fantom_cat.lv3_robust")
ref_names= c("gencode","refseq","fantom")
indexes = paste0(prefixes, "_all_exons.txt.gz")
TSSes = paste0(prefixes, ".TSS.txt")
names(indexes) = ref_names
names(TSSes) = ref_names

load_index = function(filename, path){
  index = read.delim(file.path(path, filename),
                          col.names=c("chr","exon_start","exon_end","strand","gene"))
  index = distinct(index)
  print(nrow(index))
  return(index)
}

load_TSS = function(filename, path){
  tss = read.delim(file.path(path, filename), col.names = c("chr","TSS", "strand"))
  tss = distinct(tss) 
  print(nrow(tss))
  return(tss)
}
#gencode = load_index(indexes[1], LEAF_INDEX)
#gTSS = load_TSS(TSSes[1], LEAF_INDEX)
```

```{r}
intron_to_TSS = function(first_introns, index_fn, tss_fn, path=LEAF_INDEX){
  exon_index = load_index(index_fn, path)
  tss = load_TSS(tss_fn, path)
  

  first_introns$gene = NULL
  pos_upexons = merge(exon_index, filter(first_introns,strand =="+"),  
                  by.x=c("chr", "strand", "exon_end"), 
                  by.y=c("chr", "strand", "intron_start"))

  neg_upexons = merge(exon_index, filter(first_introns,strand =="-"),  
                  by.x=c("chr", "strand", "exon_start"), 
                  by.y=c("chr", "strand", "intron_end"))

  
  first_exons = rbind(filter(pos_upexons,exon_start %in% tss$TSS),
                    filter(neg_upexons, exon_end %in% tss$TSS) %>% rename(intron_end=intron_start))
  
  print(sprintf("all exons: %d\n only TSSes: %d",nrow(pos_upexons) + nrow(neg_upexons), nrow(first_exons)))
  return(first_exons)
}
```

```{r}
gencode_TSSes = intron_to_TSS(filter(first_introns, annotation =="gencode"),
                              indexes[["gencode"]], TSSes[["gencode"]], path=LEAF_INDEX)
nrow(filter(gencode_TSSes, p.adjust < 0.05 & abs(deltapsi) > 0.1))
filter(gencode_TSSes, p.adjust < 0.05 & abs(deltapsi) > 0.1) %>% pull(intron_id) %>% n_distinct()
```


```{r}
refseq_TSSes = intron_to_TSS(filter(first_introns, annotation =="refseq"),
                              indexes[["refseq"]], TSSes[["refseq"]], path=LEAF_INDEX)
filter(refseq_TSSes,gene == "PEMT")
filter(refseq_TSSes, p.adjust < 0.05 & abs(deltapsi) > 0.1) %>% pull(intron_id) %>% n_distinct()
```


```{r}
fantom_TSSes = intron_to_TSS(filter(first_introns, annotation =="fantom_cat"),
                              indexes[["fantom"]], TSSes[["fantom"]], path=LEAF_INDEX)
filter(fantom_TSSes, p.adjust < 0.05 & abs(deltapsi) > 0.1) %>% pull(intron_id) %>% n_distinct()
```


```{r}
all_first = bind_rows(gencode_TSSes, refseq_TSSes, fantom_TSSes)
head(all_first)
table(all_first$annotation)
table(filter(all_first, p.adjust < 0.05 & abs(deltapsi)>0.1) %>% pull("annotation"))
table(filter(all_first, p.adjust < 0.05 & abs(deltapsi)>0.1) %>% select("annotation","intron_id") %>% distinct %>% pull("annotation"))
rm(gencode_TSSes, refseq_TSSes, fantom_TSSes)
```

## select background and save first exon files

```{r}
all_first = select(all_first, chr, exon_start, exon_end, intron_id, deltapsi, strand, p.adjust, gene) %>% 
    mutate(exon_start = exon_start -1) %>% #adjustment from gtf to bed coordinates  
    arrange(chr, exon_start, exon_end)
all_first = distinct(all_first)
nrow(all_first)
nrow(distinct(all_first[c("chr", "exon_start", "exon_end","strand")]))
nrow(filter(all_first, p.adjust < 0.05 & abs(deltapsi)>0.1)) #1031
nrow(distinct(filter(all_first, p.adjust < 0.05 & abs(deltapsi)>0.1) %>% select("intron_id"))) #363
head(all_first)
```

```{r}
head(all_first)
write.table(all_first, here("31_leafcutter", "histone_profile/lc_3db_all_first_exons.bed"), 
             quote=F, sep="\t", row.names = F, col.names = F)
```


## TSS only

```{r}
all_TSSes = mutate(all_first, TSS_start = if_else(strand == "+", exon_start, exon_end-1),
                    TSS_end = if_else(strand=="+", exon_start+1, exon_end)) %>% 
    arrange(chr, TSS_start, TSS_end)
head(all_TSSes)
length(unique(all_TSSes$intron_id[all_TSSes$p.adjust < 0.05 & abs(all_TSSes$deltapsi)>0.1]))
```


### Filter to 1 junction = 1 TSS
Pick the upstream TSS so we're consistent


```{r}
filter_TSS = function(strand, coords){
    if (unique(strand) == "+"){return (min(coords))
        }else if (unique(strand) == "-"){return(max(coords))}
}

tss_filt = group_by(all_TSSes, intron_id, deltapsi, p.adjust, strand, chr) %>% 
    summarise(TSS_start= filter_TSS(strand, TSS_start),  TSS_end = filter_TSS(strand, TSS_end), gene=paste(unique(gene), collapse=",")) %>%
    select(chr, TSS_start, TSS_end, intron_id, deltapsi, strand, p.adjust, gene) %>% 
    arrange(chr, TSS_start, TSS_end)

nrow(tss_filt)
head(tss_filt)
```

```{r}
nrow(filter(tss_filt, abs(deltapsi) > 0.1 & p.adjust < 0.05))


white = filter(tss_filt, deltapsi < -0.1 & p.adjust < 0.05); nrow(white)
beige = filter(tss_filt, deltapsi > 0.1 & p.adjust < 0.05); nrow(beige)
head(beige)
grep("PEMT|PPARG", white$gene, value=T)
grep("PEMT|PPARG", beige$gene, value=T)
length(unique(beige$intron_id))
```


```{r}
write.table(white, here("31_leafcutter", "histone_profile/lc_3db_white_TSSes.bed"), 
            quote=F, sep="\t", row.names = F, col.names = F)
write.table(beige, here("31_leafcutter", "histone_profile/lc_3db_beige_TSSes.bed"), 
             quote=F, sep="\t", row.names = F, col.names = F)
```

## Not-significant TSSes
### Minus DEGs
```{r}
deg = read.delim(here("03limma/any_and_all_donor_DGE.tsv"))
s6_deg = filter(deg,adj.P.Val.s6 < 0.01)
nrow(s6_deg)
is_deg = function(intron_id){
    gene = gsub("_.*","",intron_id)
    res = gene %in% s6_deg$gene_name
    return (res)
    }

nrow(tss_filt)
not_sig = filter(tss_filt, p.adjust > 0.05 & abs(deltapsi) < 0.1 & !is_deg(gene))
nrow(not_sig)#expressed, but not DSG or DEG in donor13

not_sig[grep("PEMT",not_sig$gene),]
write.table(not_sig, here("31_leafcutter", "histone_profile/lc_3db_not_sig_TSSes.bed"), 
             quote=F, sep="\t", row.names = F, col.names = F)
```


## TSS -2kb/+500b 

```{r}
promoters = mutate(.keep="none",
                   all_TSSes, chr, 
                   promoter_start=if_else(strand == "+", TSS_start-2000, TSS_end-500),
                   promoter_end =if_else(strand=="+",TSS_start+500, TSS_end+2000), 
                   intron_id, deltapsi,strand, p.adjust) %>%
    select(chr, promoter_start, promoter_end, intron_id, deltapsi,strand, p.adjust) %>% 
    arrange(chr, promoter_start, promoter_end, strand)
head(promoters)
tail(promoters)

#Convert coords to strings to avoid sci notation (bedtools won't have it)
promoters = mutate(promoters, promoter_start = format(promoter_start, scientific=F, trim=T),
                   promoter_end = format(promoter_end, scientific=F, trim=T))
```

```{r}
white_promoters = filter(promoters, p.adjust < 0.05 & deltapsi <= -0.1) 
nrow(white_promoters)
beige_promoters = filter(promoters, p.adjust < 0.05 & deltapsi >= 0.1) 
nrow(beige_promoters)
```

```{r}
write_tsv(white_promoters, here("31_leafcutter", "promoter_binding/lc_3db_white_promoters.bed"), 
         col_names=F)
write_tsv(beige_promoters, here("31_leafcutter", "promoter_binding/lc_3db_beige_promoters.bed"), 
             col_names = F)
```
