==> R/17_go_networks_plus_trifid.Rmd <==
---
title: "go_networks_plus_trifid"
output: pdf_document
date: "2023-12-15"
---
Figure 2F

```{r message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(ggplot2)
library(enrichplot)
library(clusterProfiler)
library(here); i_am("R/17_go_networks_plus_trifid.Rmd")
```


```{r}
figs =  here("R/plots")
load(file.path(figs, "Themogenesis_object.RData"))
specific = arrange(specific, desc(Count))
head(specific, n=10)[3:9]
cnetplot(specific, showCategory=6, categorySize="geneNum")
```
depicts the linkages of genes and biological concepts (e.g. GO terms or KEGG pathways) as a network. 

to plot the trifid score, we just need a sorted list of these with names of genes

```{r}
trifid = read.delim(here("31_leafcutter", "trifid_DIFFERENCE_with_Alt_introns.tsv"))
diff = trifid$trifid_diff
names(diff) = trifid$gene_to_plot
diff = diff[order(diff, decreasing=T)]
summary(diff)
```

## If multiple gene names choose

```{r}
select_gene = function(gene_str){
  genel = stringr::str_split_1(gene_str, pattern = ",")
  unknown = grepl("^A[CP}[0-9]{3,}", genel)
  if (!unknown[1]){
    return(genel[1])
  } else if (!unknown[2]){
    return(genel[2])
  } else if (!unknown[3]){
    return(genel[3])
  } else{
    print("unknown gene")
    return(genel[1])}
}

sapply(grep(",", names(diff), value=T), select_gene)

names(diff)[grep(",", names(diff))] = sapply(grep(",", names(diff), value=T), select_gene, USE.NAMES=F)
```



```{r}
shorten = function(ont) {
  no_beg = gsub("^(GO..|HALLMARK|WP|REACTOME)_", "" ,ont)
  abb = abbreviate(stringr::str_to_title(gsub("_"," ", no_beg)),minlength=40, dot=T, named = F)
  return(abb)
}

specific = mutate(specific, Description = shorten(ID) )

cnetplot(specific, showCategory=6, categorySize="geneNum", foldChange=diff)

cnetplot(specific, showCategory=5, categorySize="geneNum", color.params = list(foldChange = diff),
         color_category='gold3' ) + scale_colour_gradient2(high="red", low="cyan3", mid="black", na.value="grey", name="TRIFID_diff")


cnetplot(filter(specific,ID %in% c("GOCC_MITOCHONDRIAL_MATRIX","GOBP_ADAPTIVE_THERMOGENESIS","HALLMARK_FATTY_ACID_METABOLISM")),
         showCategory=5, categorySize="geneNum", color.params = list(foldChange = diff),
         color_category='gold3' ) + scale_colour_gradient2(high="red", low="cyan3", mid="black", na.value="grey", name="TRIFID_diff")
ggsave(file.path(figs, "thermogenesis_network_w_trifid.pdf"), width=9, height=7)
```

```{r}
heatplot(specific, foldChange = diff) + theme_classic(base_size=18) + 
  theme(axis.text.x = element_text(angle=90, vjust = 0.5, hjust=1))  + 
  scale_fill_gradient2(high="red", low="cyan3", mid="black", na.value="grey", name="TRIFID_diff")
ggsave(file.path(figs, "thermogenesis_nheatplot_w_trifid.pdf"), width=15, height=4)
```

```{r}
load(file.path(figs, "GO_object.RData"))
head(sig_ob)
sig_ob = mutate(sig_ob, Description = shorten(ID) )

cnetplot(sig_ob, showCategory=3, categorySize="geneNum", color.params = list(foldChange = diff),
         color_category='gold3' ) + scale_colour_gradient2(high="red", low="cyan3", mid="black", na.value="grey", name="TRIFID_diff")
```

```{r}
heatplot(sig_ob, foldChange = diff, showCategory = 3) + theme_classic(base_size=18) + 
  theme(axis.text.x = element_text(angle=45, hjust=1)) + guides(fill = guide_colourbar(reverse=F)) + 
  scale_fill_gradient2(high="red", low="cyan3", mid="black", na.value="grey", name="TRIFID_diff")
ggsave(file.path(figs, "go_heatplot_w_trifid.pdf"), width=15, height=3)
```
