==> R/21_pemt_cage_clusters.Rmd <==
---
title: "21_pemt_cage_clusters"
output: pdf_document
---

Supplementary Figure 8

```{r message=FALSE, warning=FALSE}
library(ComplexHeatmap)
library(circlize)
library(here); i_am("R/21_pemt_cage_clusters.Rmd")
```


```{r}
cage_counts = read.delim(here("annotations/primary_cell.hCAGE_pemt.counts.tsv"), quote = "\'")
colnames(cage_counts)
colnames(cage_counts) = gsub("\\.{1,3}","_",gsub("X.","", gsub(".CNhs.*.hg38.nobarcode.bam","", colnames(cage_counts))))

cage_counts = cage_counts[order(cage_counts$start, cage_counts$end),]
rownames(cage_counts) = paste0("c",nrow(cage_counts):1, "@PEMT")
cage_counts[,1:3] #TSS positions
```


```{r}
countm = as.matrix(cage_counts[!colnames(cage_counts) %in% c("chr","start","end")])
countm[,1:5]

#drop mesenchymal samples
countm=countm[,!grepl("Mesenchymal",colnames(countm))]
#drop doenstream PEMT
countm=countm[!grepl("PEMT@(8|9|10|11)",rownames(countm)),]

```

```{r}
samples = data.frame(sample =colnames(countm),
                     cell_type=gsub("_.*","",colnames(countm),),
                    donor=gsub(".*_","",colnames(countm)),
                    depot=gsub(".*_","",gsub("_donor.*","",colnames(countm))))
samples
colnames(countm) =paste(samples$cell_type, samples$depot)
```

```{r}
Heatmap(t(countm), cluster_columns = F, col = colorRamp2(c(0, max(countm)), c("white", "red")), name="counts")

Heatmap(t(scale(log(countm+0.5))), cluster_columns = F, col = colorRamp2(c(-2,0,2), c("blue","white", "red")),
        name="Relative\nLog\nExpression")

pdf(file = here("R/plots/pemt_cage_enrichment.pdf"))
Heatmap(t(scale(log(countm+0.5))), cluster_columns = F, col = colorRamp2(c(-2,0,2), c("blue","white", "red")),
        name="Relative\nLog\nExpression")
dev.off()
```
